+++ build.gradle
@@ -11,11 +11,7 @@ base {
 }
 
 repositories {
-	maven { url 'https://maven.shedaniel.me/' } // Cloth Config
-	maven { url 'https://maven.terraformersmc.com/releases/' } // Mod Menu
-	maven {
-		url = "https://api.modrinth.com/maven"
-	}
+	maven { url = "https://api.modrinth.com/maven" }
 }
 
 configurations.all {
@@ -43,12 +39,7 @@ dependencies {
 	modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
 	modImplementation "net.fabricmc.fabric-api:fabric-api:${project.fabric_version}"
 
-	modApi "com.terraformersmc:modmenu:${project.modmenu_version}"
 	modCompileOnly "maven.modrinth:rI0hvYcd:${project.visuality_version}"
-
-	modApi("me.shedaniel.cloth:cloth-config-fabric:${project.ccapi_version}") {
-		exclude(group: "net.fabricmc.fabric-api")
-	}
 }
 
 processResources {
