+++ gradle.properties
@@ -6,15 +6,13 @@ org.gradle.parallel=true
 # check these on https://fabricmc.net/develop
 minecraft_version=1.20.1
 yarn_mappings=1.20.1+build.10
-loader_version=0.15.6
+loader_version=0.16.7
 
 # Mod Properties
-mod_version=1.4.0
-maven_group=com.juicey.juiceytweaks
+mod_version=1.5.1
+maven_group=com.juiceybeans.juiceytweaks
 archives_base_name=juiceytweaks
 
 # Dependencies
-fabric_version=0.91.0+1.20.1
-modmenu_version=7.1.0
+fabric_version=0.92.2+1.20.1
 visuality_version=uhvQD6Ny
-ccapi_version=11.1.106
