--- src/main/java/com/juicey/juiceytweaks/integration/ModMenuIntegration.java
@@ -1,18 +1,0 @@
-package com.juicey.juiceytweaks.integration;
-
-import com.terraformersmc.modmenu.api.ConfigScreenFactory;
-import com.terraformersmc.modmenu.api.ModMenuApi;
-import me.shedaniel.autoconfig.AutoConfig;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
-import com.juicey.juiceytweaks.config.ModConfig;
-
-@Environment(EnvType.CLIENT)
-public class ModMenuIntegration implements ModMenuApi
-{
-    /*@Override
-    public ConfigScreenFactory<?> getModConfigScreenFactory()
-    {
-        return parent -> AutoConfig.getConfigScreen(ModConfig.class, parent).get();
-    }*/
-}
