+++ src/main/java/com/juiceybeans/juiceytweaks/item/ModItems.java
@@ -1,6 +1,6 @@
-package com.juicey.juiceytweaks.item;
+package com.juiceybeans.juiceytweaks.item;
 
-import com.juicey.juiceytweaks.JuiceyTweaks;
+import com.juiceybeans.juiceytweaks.JuiceyTweaks;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.minecraft.item.Item;
@@ -18,14 +18,13 @@ public class ModItems {
         return Registry.register(Registries.ITEM, new Identifier(JuiceyTweaks.MOD_ID, name), item);
     }
     public static final Item BLANK_SMITHING_TEMPLATE = registerItem("blank_smithing_template", new Item(new FabricItemSettings()));
-    public static final Item ELDER_SPIKE = registerItem("elder_spike", new Item(new FabricItemSettings()));
 
     public static void addItemsToItemGroup() {
         //For future reference:
         //addToItemGroup(ItemGroups.[item group], [mod item]);
 
-        ItemGroupEvents.modifyEntriesEvent(ItemGroups.INGREDIENTS).register(entries -> entries.addAfter(Items.NETHERITE_UPGRADE_SMITHING_TEMPLATE, ModItems.BLANK_SMITHING_TEMPLATE));
-        addToItemGroup(ItemGroups.INGREDIENTS, ELDER_SPIKE);
+        ItemGroupEvents.modifyEntriesEvent(ItemGroups.INGREDIENTS).register(entries -> entries.addAfter(
+                Items.NETHERITE_UPGRADE_SMITHING_TEMPLATE, ModItems.BLANK_SMITHING_TEMPLATE));
     }
 
     public static void addToItemGroup(RegistryKey<ItemGroup> group, Item item) {
