+++ src/main/java/com/juiceybeans/juiceytweaks/mixin/PowderSnowMixin.java
@@ -1,8 +1,6 @@
-package com.juicey.juiceytweaks.mixin;
+package com.juiceybeans.juiceytweaks.mixin;
 
-import com.juicey.juiceytweaks.JuiceyTweaks;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
-import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import net.minecraft.block.PowderSnowBlock;
 import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.enchantment.Enchantments;
