+++ src/main/java/com/juiceybeans/juiceytweaks/mixin/SculkCatalystMixin.java
@@ -1,4 +1,4 @@
-package com.juicey.juiceytweaks.mixin;
+package com.juiceybeans.juiceytweaks.mixin;
 
 import net.minecraft.block.*;
 import net.minecraft.particle.ParticleTypes;
@@ -10,9 +10,9 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(SculkVeinBlock.class)
-public abstract class SculkVeinMixin extends MultifaceGrowthBlock {
-    public SculkVeinMixin(Settings settings) {
+@Mixin(SculkCatalystBlock.class)
+public abstract class SculkCatalystMixin extends BlockWithEntity {
+    public SculkCatalystMixin(AbstractBlock.Settings settings) {
         super(settings);
     }
 
@@ -23,13 +23,14 @@ public abstract class SculkVeinMixin extends MultifaceGrowthBlock {
     @Override
     public void randomDisplayTick(BlockState state, World world, BlockPos pos, Random random) {
         super.randomDisplayTick(state, world, pos, random);
-        if(state.getBlock() instanceof SculkVeinBlock && random.nextFloat() > 0.5) {
+        if(state.getBlock() instanceof SculkCatalystBlock && random.nextFloat() > 0.5) {
             double x = pos.getX() + random.nextDouble();
             double y = pos.getY() + random.nextDouble();
             double z = pos.getZ() + random.nextDouble();
 
             if (random.nextFloat() > 0.8) {
                 world.addParticle(ParticleTypes.SCULK_CHARGE_POP, x + 0.5, y + 0.5, z + 0.5, 0, 0, 0);
+                world.addParticle(ParticleTypes.SCULK_SOUL, x + 0.5, y + 0.5, z + 0.5, 0, 0, 0);
             }
         }
