+++ src/main/java/com/juiceybeans/juiceytweaks/mixin/SculkMixin.java
@@ -1,4 +1,4 @@
-package com.juicey.juiceytweaks.mixin;
+package com.juiceybeans.juiceytweaks.mixin;
 
 import net.minecraft.block.*;
 import net.minecraft.particle.ParticleTypes;
@@ -10,9 +10,9 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(SculkShriekerBlock.class)
-public abstract class SculkShriekerMixin extends BlockWithEntity {
-    public SculkShriekerMixin(AbstractBlock.Settings settings) {
+@Mixin(SculkBlock.class)
+public class SculkMixin extends ExperienceDroppingBlock {
+    public SculkMixin(Settings settings) {
         super(settings);
     }
 
@@ -23,13 +23,13 @@ public abstract class SculkShriekerMixin extends BlockWithEntity {
     @Override
     public void randomDisplayTick(BlockState state, World world, BlockPos pos, Random random) {
         super.randomDisplayTick(state, world, pos, random);
-        if(state.getBlock() instanceof SculkShriekerBlock && random.nextFloat() > 0.5) {
+        if(state.getBlock() instanceof SculkBlock && random.nextFloat() > 0.5) {
             double x = pos.getX() + random.nextDouble();
             double y = pos.getY() + random.nextDouble();
             double z = pos.getZ() + random.nextDouble();
 
             if (random.nextFloat() > 0.8) {
-                world.addParticle(ParticleTypes.SCULK_SOUL, x + 0.5, y + 0.5, z + 0.5, 0, 0, 0);
+                world.addParticle(ParticleTypes.SCULK_CHARGE_POP, x + 0.5, y + 0.5, z + 0.5, 0, 0, 0);
             }
         }
