+++ src/main/java/com/juiceybeans/juiceytweaks/mixin/SculkSensorMixin.java
@@ -1,6 +1,7 @@
-package com.juicey.juiceytweaks.mixin;
+package com.juiceybeans.juiceytweaks.mixin;
 
 import net.minecraft.block.*;
+import net.minecraft.particle.ParticleEffect;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.random.Random;
@@ -10,9 +11,9 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(SculkBlock.class)
-public class SculkMixin extends ExperienceDroppingBlock {
-    public SculkMixin(Settings settings) {
+@Mixin(SculkSensorBlock.class)
+public abstract class SculkSensorMixin extends BlockWithEntity {
+    public SculkSensorMixin(AbstractBlock.Settings settings) {
         super(settings);
     }
 
@@ -23,12 +24,12 @@ public class SculkMixin extends ExperienceDroppingBlock {
     @Override
     public void randomDisplayTick(BlockState state, World world, BlockPos pos, Random random) {
         super.randomDisplayTick(state, world, pos, random);
-        if(state.getBlock() instanceof SculkBlock && random.nextFloat() > 0.5) {
+        if(state.getBlock() instanceof SculkSensorBlock && random.nextFloat() > 0.5) {
             double x = pos.getX() + random.nextDouble();
             double y = pos.getY() + random.nextDouble();
             double z = pos.getZ() + random.nextDouble();
 
-            if (random.nextFloat() > 0.8) {
+            if (random.nextFloat() > 0.85) {
                 world.addParticle(ParticleTypes.SCULK_CHARGE_POP, x + 0.5, y + 0.5, z + 0.5, 0, 0, 0);
             }
         }
