+++ src/main/resources/assets/juiceytweaks/lang/en_us.json
@@ -1,5 +1,4 @@
 {
   "item.juiceytweaks.blank_smithing_template": "Blank Smithing Template",
-  "item.juiceytweaks.elder_spike": "Elder Spike",
-  "itemGroup.juiceytweaks.JUICEY_TWEAKS": "Juicey Tweaks"
+  "itemGroup.juiceytweaks.juicey_tweaks": "Juicey's Tweaks"
 }
