--- src/main/resources/assets/juiceytweaks/models/item/elder_spike.json
@@ -1,6 +1,0 @@
-{
-  "parent": "item/generated",
-  "textures": {
-    "layer0": "juiceytweaks:item/elder_spike"
-  }
-}
