--- src/main/resources/assets/juiceytweaks/models/item/elder_spike_in_hand.json
@@ -1,22 +1,0 @@
-{
-  "parent": "builtin/entity",
-  "credit": "Made with Blockbench",
-  "textures": {
-    "0": "elder_spike_hand",
-    "particle": "elder_spike_hand"
-  },
-  "elements": [
-    {
-      "from": [7, 0, 7],
-      "to": [9, 9, 9],
-      "faces": {
-        "north": {"uv": [0, 3, 4, 16], "texture": "#0"},
-        "east": {"uv": [12, 3, 16, 16], "texture": "#0"},
-        "south": {"uv": [8, 3, 12, 16], "texture": "#0"},
-        "west": {"uv": [4, 3, 8, 16], "texture": "#0"},
-        "up": {"uv": [4, 0, 8, 3], "texture": "#0"},
-        "down": {"uv": [8, 0, 12, 3], "texture": "#0"}
-      }
-    }
-  ]
-}
