--- src/main/resources/data/minecraft/recipes/cookie.json
@@ -1,20 +1,0 @@
-{
-  "type": "minecraft:crafting_shaped",
-  "category": "misc",
-  "key": {
-    "#": {
-      "item": "minecraft:wheat"
-    },
-    "X": {
-      "item": "minecraft:cocoa_beans"
-    }
-  },
-  "pattern": [
-    "#X#"
-  ],
-  "result": {
-    "count": 3,
-    "item": "minecraft:cookie"
-  },
-  "show_notification": true
-}
