+++ src/main/resources/fabric.mod.json
@@ -1,14 +1,13 @@
 {
 	"schemaVersion": 1,
 	"id": "juiceytweaks",
-	"version": "1.4.0",
+	"version": "1.5.1",
 	"name": "Juicey's Tweaks",
 	"description": "Glowing torchflowers, freezing snowballs, better trim recipes, and more!",
 	"authors": [
 		"JuiceyBeans"
 	],
 	"contact": {
-		"homepage": "https://discord.gg/N7cJ367Ngs",
 		"sources": "https://github.com/JuiceyBeans/JuiceyTweaks"
 	},
 	"license": "MIT",
@@ -16,10 +15,7 @@
 	"environment": "*",
 	"entrypoints": {
 		"main": [
-			"com.juicey.juiceytweaks.JuiceyTweaks"
-		],
-		"modmenu": [
-			"com.juicey.juiceytweaks.integration.ModMenuIntegration"
+			"com.juiceybeans.juiceytweaks.JuiceyTweaks"
 		]
 	},
 	"mixins": [
@@ -27,14 +23,12 @@
 	],
 	"depends": {
 		"fabricloader": ">=0.15.0",
-		"minecraft": "~1.20.1",
+		"minecraft": "~1.21.2",
 		"java": ">=17",
 		"fabric-api": "*",
 		"cloth-config": "*"
 	},
 	"suggests": {
-		"modmenu": "*",
-		"autoconfig1u": "*",
 		"visuality": "*"
 	}
 }
